pyproject.toml
[project]
name = "maker"
version = "0.1.0"
description = "CLI tool to make stuff."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "yt-dlp>=2023.0.0",
    "ffmpeg-python>=0.2.0",
    "pydantic[email]>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
maker = "maker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.14.11"]
